LF-11115 imx: ahab: Use authenticated header for images loading
When authenticating container header, the container header is first
loaded to the heap memory in DRAM and then copied to a different
location in RAM for authentication (in the function ahab_auth_cntr_hdr).
To mitigate risk, the container header being authenticated should be
the one that is parsed by the loader SW.
The patch refactors the ahab_auth_cntr_hdr to return the container
be authenticated. Caller uses this authenticated container for following
parsing and image loading.

Signed-off-by: Ye Li <ye.li@nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>