LFU-739-2 imx91_evk: Workaround LCD panel power leakage
iMX91 has power leakage on pins GPIO0/1/2/3, the pins are muxed as
CLK/DE/VSYNC/HSYNC to LCD panel. The leakage to LCD panel causes power
leakage on vRPi 3v3 and affects LCD panel sometime.

To work around the issue, when the panel is not working, we have to set
output low for GPIO0/1/2/3 to avoid leakage. So SPL has to set them as
early as possible.

Turn off the vRPi 5V and RPi EXP GPIO mux selection in u-boot
board_quiece_devices, so the kernel can use its sequence to avoid
backlight turned on (white screen) before LCDIF data arriving.

And set PWM pin to high to disable backlight as default in u-boot,
otherwise backlight will immediately turn on when 5V is enabled and
shows white screen.

Signed-off-by: Ye Li <ye.li@nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>