MA-23254 force assign secure memory to linux
Because widevine L1 is enabled, linux and Android will use secure heap
and vpu boot memory, and the corresponding virtual address will be
mapped in the linux kernel, so these two sections of memory must be
passed to linux so that the memory nodes in dts contain them. Otherwise,
it will dump during linux map.

Change-Id: I7d4da58c28b60e1e486900ab88dffd0f0d518c7e
Signed-off-by: He Zhai <zhai.he@nxp.com>