MA-23203 Encapsulate the bootloader rollback index as a blob
To avoid disclosure of plaintext data in rpmb, this encapsulates
the bootloader rollback index as a blob then keep in rpmb.
set "CONFIG_IMX_ROLLBACK_BLOB=y" in uboot configuration file to
enable this function.

Also, the RPMB magic number is storaged as plaintext in rpmb. To
avoid hacker modify the RPMB magic number, if "CONFIG_IMX_ROLLBACK_BLOB=y"
be set, boot up will be denied if the RPMB magic number is not
matched.
To initialize the RPMB magic number, set "CONFIG_RPMB_INIT_ALLOW=y"
in uboot configuration file, This allow boot and RPMB initializing
even though the RPMB magic not matched. This config should be set
for production only and disabled for release.

This function not be enabled on i.MX 9 products due to ele not
supported yet.

Change-Id: I7a1501f04b3d78f1de7a996911c108d45d8168f4
Signed-off-by: Maximus <maximus.sun@nxp.com>