LFU-827 lx2160/lx2162: Fix incorrect CONFIG_SYS_FSL_NUM_CC_PLLS
When configuring A72 CC5 to CC8 to use CGB_PLL2, u-boot displayed
clock configuration of CPU9-15 as 0 MHz.

The root cause is the structures used in get_sys_info function
always assume 3 PLLs in each CGn, even it is 2 PLL in each on lx2160.
Setting CONFIG_SYS_FSL_NUM_CC_PLLS to 4 causing calculation problem.

Users should set it to 5 or 6 to fix the issue.

Signed-off-by: Ye Li <ye.li@nxp.com>
Tested-by: Ran Tang <ran.tang@nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>