LFU-898-7 spl: spl_imx_container: Support loading for recovery image
The container used by CRRM will include Normal OS and Recovery OS
(with initramfs). In imx-mkimage, a new image type "recovery" is added to
indicate the type. When loading container, according to "recovery" field
in spl_image to determine if need to load Recovery OS or not. User has to
override arch_spl_load_configure to set recovery field.

Signed-off-by: Ye Li <ye.li@nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>